requirements.txt
django
channels
daphne
redis
channels-redis
numpy
pandas
django_rest_framework
python-decouple
gunicorn
